ci: automate semantic version tag increment
On any update to main (commit or pr), the semantic version is automatically updated.

If no bump type is specified in the commit message, the patch is incremented (i.e., v0.0.X).

If `(minor)` is included in the commit message, the minor version is incremented (i.e., v0.X.0).

If `(major)` is included in the commit message, the major version is incremented (i.e., vX.0.0).